Move source to source directory
