--- PlayniteVndbExtension/VndbMetadataSettingsView.xaml
@@ -1,91 +1,0 @@
-﻿<UserControl x:Class="VndbMetadata.VndbMetadataSettingsView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:sys="clr-namespace:System;assembly=mscorlib"
-             xmlns:enums="clr-namespace:VndbMetadata.Models;assembly=VndbMetadata"
-             mc:Ignorable="d"
-             d:DesignHeight="400" d:DesignWidth="600">
-
-
-    <UserControl.Resources>
-        <ObjectDataProvider x:Key="GetSpoilerLevelsFromEnum" MethodName="GetValues" ObjectType="{x:Type sys:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="enums:SpoilerLevel" />
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <ObjectDataProvider x:Key="GetViolenceLevelsFromEnum" MethodName="GetValues" ObjectType="{x:Type sys:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="enums:ViolenceLevel" />
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <ObjectDataProvider x:Key="GetSexualityLevelsFromEnum" MethodName="GetValues" ObjectType="{x:Type sys:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="enums:SexualityLevel" />
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-    </UserControl.Resources>
-    
-    <StackPanel Margin="20">
-        <TextBlock HorizontalAlignment="Left" FontWeight="Bold" Text="Images:" />
-        <StackPanel Orientation="Horizontal" Margin="10">
-            <StackPanel Margin="5">
-                <TextBlock Text="Maximum Violence Level" HorizontalAlignment="Center" />
-                <ComboBox SelectedValue="{Binding Settings.ImageMaxViolenceLevel}" MinWidth="100"
-                          ItemsSource="{Binding Source={StaticResource GetViolenceLevelsFromEnum}}" />
-            </StackPanel>
-            <StackPanel Margin="5">
-                <TextBlock Text="Maximum Sexuality Level" HorizontalAlignment="Center" />
-                <ComboBox SelectedValue="{Binding Settings.ImageMaxSexualityLevel}" MinWidth="100"
-                          ItemsSource="{Binding Source={StaticResource GetSexualityLevelsFromEnum}}" />
-            </StackPanel>
-        </StackPanel>
-        <Separator />
-        <TextBlock HorizontalAlignment="Left" FontWeight="Bold" Text="Tags:" />
-        <StackPanel Orientation="Horizontal" Margin="5">
-            <StackPanel Margin="5">
-                <TextBlock Text="Max Tags" TextAlignment="Center" MinWidth="320"/>
-                <TextBox HorizontalAlignment="Center" Text="{Binding Settings.MaxAllTags}" MinWidth="30" />
-            </StackPanel>
-        </StackPanel>
-        <StackPanel Orientation="Horizontal" Margin="5">
-            <StackPanel Margin="5">
-                <TextBlock Text="Max Content Tags" TextAlignment="Center" MinWidth="100"/>
-                <TextBox HorizontalAlignment="Center" Text="{Binding Settings.MaxContentTags}" MinWidth="30"></TextBox>
-            </StackPanel>
-            <StackPanel Margin="5">
-                <TextBlock Text="Max Sexual Tags" TextAlignment="Center" MinWidth="100" />
-                <TextBox HorizontalAlignment="Center" Text="{Binding Settings.MaxSexualTags}" MinWidth="30"></TextBox>
-            </StackPanel>
-            <StackPanel Margin="5">
-                <TextBlock Text="Max Technical Tags" TextAlignment="Center" MinWidth="100" />
-                <TextBox HorizontalAlignment="Center" Text="{Binding Settings.MaxTechnicalTags}" MinWidth="30"></TextBox>
-            </StackPanel>
-        </StackPanel>
-        <StackPanel Orientation="Horizontal" Margin="5">
-            <StackPanel Margin="5">
-                <TextBlock Text="Minimum Score" HorizontalAlignment="Center" />
-                <Slider Minimum="1" Maximum="3" TickFrequency="0.1" Value="{Binding Settings.TagMinScore}"
-                        AutoToolTipPlacement="BottomRight" AutoToolTipPrecision="1" MinWidth="100" />
-            </StackPanel>
-            <StackPanel Margin="5">
-                <TextBlock Text="Spoiler Level" HorizontalAlignment="Center" />
-                <ComboBox SelectedValue="{Binding Settings.TagMaxSpoilerLevel}" MinWidth="100"
-                          ItemsSource="{Binding Source={StaticResource GetSpoilerLevelsFromEnum}}" />
-            </StackPanel>
-        </StackPanel>
-        <StackPanel Orientation="Horizontal" Margin="5">
-            <TextBlock Text="Tags last Update on " />
-            <TextBlock x:Name="TagDateField" Text="{Binding Settings.LastTagUpdate }" MinWidth="100" />
-        </StackPanel>
-        <Separator />
-        <StackPanel>
-            <TextBlock HorizontalAlignment="Left" FontWeight="Bold" Text="Dates:" />
-            <StackPanel Orientation="Horizontal" Margin="10">
-                <TextBlock Text="Allow incomplete dates: " />
-                <CheckBox IsChecked="{Binding Settings.AllowIncompleteDates}" />
-            </StackPanel>
-        </StackPanel>
-    </StackPanel>
-</UserControl>
