--- PlayniteVndbExtension/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net462" />
-  <package id="PlayniteSDK" version="6.1.0" targetFramework="net462" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net462" />
-  <package id="VndbSharp" version="1.1.1" targetFramework="net462" />
-</packages>
