+++ source/VndbMetadata.csproj
